requirements.txt
# Telegram Bot
python-telegram-bot==21.8

# OCR
pytesseract==0.3.13
Pillow>=10.3.0

# Excel y datos
pandas>=2.2.0
openpyxl>=3.1.2

# Variables de entorno
python-dotenv>=1.0.0